+++ inc/functions/setSessionLinkToBuyTicket.php
@@ -1,0 +1,4 @@
+<?php
+function setSessionLinkToBuyTicket(&$session, $link) {
+    $session["linkToBuyTicket"] = $link;
+}