+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2023-03-01
+
+### Added
+
+- Button to add a ticket directly in the session
+- Custom link to buy ticket per session
+
+### Fixed
+
+- Check duplication when saving/fetching session players
+
 ## [1.4.0] - 2023-05-05
 
 ### Added
@@ -36,7 +47,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial version
 
-[unreleased]: https://github.com/neolao/soccer-session/compare/1.4.0...HEAD
+[unreleased]: https://github.com/neolao/soccer-session/compare/1.5.0...HEAD
+[1.5.0]: https://github.com/neolao/soccer-session/compare/1.4.0...1.5.0
 [1.4.0]: https://github.com/neolao/soccer-session/compare/1.3.0...1.4.0
 [1.3.0]: https://github.com/neolao/soccer-session/compare/1.2.0...1.3.0
 [1.2.0]: https://github.com/neolao/soccer-session/compare/1.1.0...1.2.0